% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie_encryption.R
\name{encrypt_cookie}
\alias{encrypt_cookie}
\alias{decrypt_cookie}
\title{Encrypt cookie.}
\usage{
encrypt_cookie(cookie_key, message)

decrypt_cookie(cookie_key, message)
}
\arguments{
\item{cookie_key}{encryption key.}

\item{message}{string to decrypt.}
}
\description{
Encrypt cookie.

Decrypt cookie.
}
