% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emayili_emailer.R
\name{emayili_emailer}
\alias{emayili_emailer}
\title{Returns a function to send emails using the \code{emayili} package.}
\usage{
emayili_emailer(
  email_host = NULL,
  email_port = NULL,
  email_username = NULL,
  email_password = NULL,
  from_email = NULL
)
}
\arguments{
\item{email_host}{SMTP email host.}

\item{email_port}{SMPT email port.}

\item{email_username}{username for the SMTP server.}

\item{email_password}{password for the SMTP server.}

\item{from_email}{the from email address sent from \code{\link[=login_server]{login_server()}}.}
}
\value{
returns a function to send an email using the \code{emayili} package.
}
\description{
This returns a function that can be used with the \code{\link[=login_server]{login_server()}}. Specifically,
the function takes two parameters, \code{to_email} and \code{message}.
}
\seealso{
\code{\link[=login_server]{login_server()}}
}
