% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_in.R
\name{is_logged_in}
\alias{is_logged_in}
\title{Display Shiny elements only if the user is logged in.}
\usage{
is_logged_in(id, ...)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}

\item{...}{Shiny UI elements.}
}
\value{
a \code{\link[shiny:conditionalPanel]{shiny::conditionalPanel()}} object.
}
\description{
This function can be used on the Shiny UI side. It will check to see if the user
is logged in, if so the other Shiny elements will be displayed.
}
\seealso{
\code{\link[=login_server]{login_server()}}
}
