% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_in.R
\name{is_not_logged_in}
\alias{is_not_logged_in}
\title{Display Shiny elements only if the user is not logged in.}
\usage{
is_not_logged_in(id, ...)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}

\item{...}{Shiny UI elements.}
}
\value{
a [@shiny::conditionalPanel()]
}
\description{
This function can be used on teh Shiny UI side. It will. check to see if the
user is not logged in, if so the other Shiny elements will be displayed.
}
\examples{
\dontrun{
login::is_not_logged_in(id = 'my_app_id',
    div('Here is some HTML code that only shows when NOT logged in.'))
}
}
