% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login-package.R
\docType{package}
\name{login-package}
\alias{login}
\alias{login-package}
\title{Shiny Login Module.}
\description{
This script is modified from Huidong Tian's
original script. The blog post describing the method is here:
http://withr.me/authentication-of-shiny-server-application-using-a-simple-method/
The original R script is located here: https://gist.github.com/withr/9001831
}
\details{
This script adds two new features: 1. Render a logout button, and 2. provide
the ability for visitors to create a new account.
}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{internal}
