% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_dependencies.R
\name{login_dependencies}
\alias{login_dependencies}
\title{JavaScript and CSS dependencies.}
\usage{
login_dependencies()
}
\value{
a [htmltools::htmlDependency()] object defining the JavaScript and CSS files.
}
\description{
JavaScript and CSS dependencies.
}
