% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_server.R
\name{login_server}
\alias{login_server}
\title{Login server module.}
\usage{
login_server(id, db_conn = NULL, users_table = "users")
}
\arguments{
\item{id}{unique ID for the Shiny Login module.}
}
\description{
Login server module.
}
