% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_ui.R
\name{login_ui}
\alias{login_ui}
\title{Login UI elements.}
\usage{
login_ui(id)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}
}
\description{
This will render (if the user is not logged in) text boxes and buttons for
the user to login.
}
