% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ui.R
\name{logout_button}
\alias{logout_button}
\title{Logout button.}
\usage{
logout_button(
  id,
  label = "Logout",
  icon = shiny::icon("right-from-bracket"),
  style = ""
)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}

\item{label}{label of the logout button.}

\item{icon}{icon for the logout button.}

\item{style}{CSS styles for the logout button.}
}
\description{
Logout button.
}
