% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ui.R
\name{logout_ui}
\alias{logout_ui}
\title{Logout button.}
\usage{
logout_ui(id, label = "Logout", icon = icon("user"), style = "")
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}
}
\description{
Logout button.
}
