% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_user_ui.R
\name{new_user_ui}
\alias{new_user_ui}
\title{UI for creating a new user account.}
\usage{
new_user_ui(id)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}
}
\description{
UI for creating a new user account.
}
