% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwdInput.R
\name{passwdInput}
\alias{passwdInput}
\title{Password input textbox.}
\usage{
passwdInput(inputId, label, value)
}
\arguments{
\item{inputId}{ID for the input.}

\item{label}{label for the textbox.}

\item{value}{default value.}
}
\value{
a [shiny::tagList()] object.
}
\description{
This is an extension to Shiny's built in passwordInput by encrpting the
password client side before sending it to the server. Although it is encrypted
in the client using JavaScript it highly recommend that you also use an SSL
certificate (for https) as well.
}
