% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_param_module.R
\name{simple_parameter_validator}
\alias{simple_parameter_validator}
\title{A simple validation check to ensure the parameters are not blank.}
\usage{
simple_parameter_validator(values, types)
}
\arguments{
\item{values}{a named list of values to check.}

\item{types}{a character vector of the input types.}
}
\value{
either \code{TRUE} if the validation passes or a character string indicating why the
validation failed.
}
\description{
A simple validation check to ensure the parameters are not blank.
}
