% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_login.R
\name{use_login}
\alias{use_login}
\title{JavaScript and CSS dependencies.}
\usage{
use_login()
}
\value{
a [htmltools::htmlDependency()] object defining the JavaScript and CSS files.
}
\description{
This ensures the JavaScript and CSS dependencies are available to the
client. Files are located in `assets/` folder when installed..
}
