% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_param_module.R
\name{showParamButton}
\alias{showParamButton}
\alias{clearParamButton}
\title{Button to show the parameter modal dialog box.}
\usage{
showParamButton(id, label = "Edit Parameters", icon = shiny::icon("gear"))

clearParamButton(id, label = "Clear Parameters", icon = shiny::icon("eraser"))
}
\arguments{
\item{id}{unique ID for the Shiny Login module.}

\item{label}{The contents of the button or link–usually a text label, but you could also use any
other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{shiny::icon()}} to appear on the button.}
}
\description{
Button to show the parameter modal dialog box.

Action button that will clear the parameters and remove cookies.
}
