% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emayili_emailer.R
\name{emayili_emailer}
\alias{emayili_emailer}
\title{Returns a function to send emails using the `emayili` package.}
\usage{
emayili_emailer(
  email_host = NULL,
  email_port = NULL,
  email_username = NULL,
  email_password = NULL,
  from_email = NULL,
  new_account_subject = "Verify your new account",
  reset_password_subject = "Reset password"
)
}
\arguments{
\item{email_host}{SMTP email host.}

\item{email_port}{SMPT email port.}

\item{email_username}{username for the SMTP server.}

\item{email_password}{password for the SMTP server.}

\item{from_email}{the from email address sent from [login_server()].}

\item{new_account_subject}{the subject used for verifying new accounts.}

\item{reset_password_subject}{the subject of password reset emails.}
}
\description{
This returns a function that can be used with the [login_server()]. Specfically,
the function takes two parameters, `to_email` and `message`.
}
\seealso{
[login_server()]
}
