% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_in.R
\name{is_not_logged_in}
\alias{is_not_logged_in}
\title{Display Shiny elements only if the user is not logged in.}
\usage{
is_not_logged_in(id, ...)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}

\item{...}{Shiny UI elements.}
}
\value{
a [@shiny::conditionalPanel()]
}
\description{
Display Shiny elements only if the user is not logged in.
}
