% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login-package.R
\docType{package}
\name{login-package}
\alias{login}
\alias{login-package}
\title{Shiny Login Module.}
\description{
This package provides a Shiny module for adding user authentication to
applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbryer/login}
  \item \url{https://jbryer.github.io/login/}
  \item Report bugs at \url{https://github.com/jbryer/login/issues}
}

}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{internal}
