% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login-package.R
\docType{package}
\name{login-package}
\alias{login}
\alias{login-package}
\title{Shiny Login Module.}
\description{
This R package was modified from Huidong Tian's
original script located here: https://gist.github.com/withr/9001831
}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{internal}
