% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_ui.R
\name{login_ui}
\alias{login_ui}
\title{Login UI elements.}
\usage{
login_ui(id, label = "Login", icon = shiny::icon("right-to-bracket"))
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}
}
\description{
Login UI elements.
}
