% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwdInput.R
\name{passwdInput}
\alias{passwdInput}
\title{Password input textbox}
\usage{
passwdInput(inputId, label, value)
}
\arguments{
\item{inputId}{ID for the input.}

\item{label}{label for the textbox.}

\item{value}{default value.}
}
\description{
Password input textbox
}
