% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_password_ui.R
\name{reset_password_ui}
\alias{reset_password_ui}
\title{UI for resetting password.}
\usage{
reset_password_ui(id)
}
\arguments{
\item{id}{id unique ID for the Shiny Login module.}
}
\description{
UI for resetting password.
}
